[Support] Add `M::Config`. (Reland #17623) (#17714)
This patch adds a simple Config structure. This structure contains a parser for an ini-like file format, and is overrideable with environment variables.

The config structure has a canonical location and name, both of which are defined in `Config::open` - this makes it easy to find, and easy for tools to write. Tools can also set the values and then flush the config to disk.

MODULAR_ORIG_COMMIT_REV_ID: 3a0dfacfb5c12cfbb4042909b2d929f6005638a5