[Reland][Kernels] Trim down the FC kernel, NFC (#17464)
This trims down the FC kernel to just contain enough for testing and
benchmarking. This also removes the epilogue function dispatch from the
matmul shim.

Pretty soon this will be moved to Faux to remove a dep.

Originally the issue is that this broke main, but it appears the issue
is resolved

Reverts [Internal link]

NN_ORIG_REV_ID: 389a5361537f8149517b961d96575cd00be7b0e2