[Kernels] Move the MHA kernels into Faux for benchmarking (#17896)
These kernels are currently only used for benchmarking and are not
exersized by mo, so we can move them to Faux to break some dependencies.
To do that we shuffled some modules around and added some options to
CMakeLists.

This also removes some tests (which I am not happy about), but I need to
think how to restore the tests and/or use the kernel-verify to validate
against a golden solution.

This is a first step in a more broad effort to take out the shims and
use them only for benchmarking and validation.

MODULAR_ORIG_COMMIT_REV_ID: dcd5286e9bb118e13285a3b455bd29028ac22cab