[KGEN] Add JITDylib-specific lookup to ExecutionEngine. (#17890)
This patch adds a version of ExecutionEngine::lookup that only looks up a symbol in a given JITDylib. This shouldn't be used generally, but when we know where the symbol is located it's useful for narrowing the search scope.

MODULAR_ORIG_COMMIT_REV_ID: f0deaf525d9532c86303e60906390264c62f195d