[KGEN] Pipe sanitizer libs into the ExecutionEngine. (#17916)
This patch uses the config machinery to allow the user to pass in a path to the sanitizer libs to use to the execution engine. If no path is provided through the config machinery, we attempt to find it using the system clang and --print-file-name.

MODULAR_ORIG_COMMIT_REV_ID: 9557cabb33bd2526b0793cd735f7edfbcb21b67e