[Telemetry] Add basic resource attributes. (#17986)
This patch adds the easy-to-get resource attributes that we want to attach to every log message, and propagates them to the OTel machinery so it's hooked in properly.

In the future, we may want to expose a way to set Resources on individual messages, but based on the list of required telemetry fields, it doesn't look like we'll need that.

MODULAR_ORIG_COMMIT_REV_ID: 0610da2f4142a72f9cbb5874fe5bd9cec5e3765d