[LSP] Index arguments (#17724)
This was a bit tricky, because there's no such thing like a well defined ArgumentDeclOp that I can use to query things. It is possible to recover the BlockArgument from the IR for all cases except owned arguments, for which there's an issue [Internal link]
I made several changes in the code to make this work, especially an improvement in MojoASTDeclRef::getName() which now falls back to the defining operator of the IR to get a name. This worked for displaying owned arguments on hover.

MODULAR_ORIG_COMMIT_REV_ID: c5af6cc5b352c6561c665acd254b8992858043d3