[LLCL] Use composition for the TelemetryContext. (#18004)
This patch allows the user to construct their own TelemetryContext and pass it into the LLCL::Runtime object. If the user opts not to, then a default context is created. This patch also moves the telemetry context out of a member variable and into the list of type-erased context objects. This should help us when it comes time to move the telemetry context back out of LLCL in the future.

The main reason to construct a context manually is to pass in your own 'resources' - which are essentially attributes that are added to each log message.

MODULAR_ORIG_COMMIT_REV_ID: 10f17af1c952790de72b981b2df4695ddf380ac2