[Support][DebugInfo][KGEN] Introduce an `InlinedSubprogramScoped` interface (#18047)
Since closure staging ops both act as a function and a callsite to that
function, they need to carry location information for both. This patch
introduces an abstraction for this, along with a utility to ease
updating locations during inlining. As a consequence, the location
handling logic is updated in the elaborator and the `force-inline` pass;
the tests for the latter are also simplified and cleaned up to make it
more readable.

MODULAR_ORIG_COMMIT_REV_ID: 5e660908b83b514c5148b5f9dbe5a7244b9b3cd0