[Support] Add Config::getValuesInSection. (#18090)
This patch adds the ability to specify a section and then list all the values in that section. This is useful for cases where we want to use the config file to (for example) list all the installed packages.

MODULAR_ORIG_COMMIT_REV_ID: 45d489c93fc898ff692671f6c36fa718999c35b8