[mojo-lang] Erase unreachable decls coming out of the parser (#18117)
Erase all declarations unreachable from the main module at the end of
parsing.
This is primarily
useful when there are imported modules that were not lazily loaded from
the
cache. This function should run after textual modules are saved to the
cache, reducing the amount of IR reaching the compiler. In addition, it
provides the compiler a canonical form of IR coming out of the parser.
This,
for instance, ensures the cache key computed on parser output does not
depend on whether the parser has cache hits for lazy loading.

MODULAR_ORIG_COMMIT_REV_ID: 1beb8cdb5d25d819fcccfa07b461727d1d5c4a11