[Kernels] Adds a 4xint8 to int32 dot product function for AVX2 and VNNI (#18124)
Adds a 4xint8 to int32 dot product function, dot_i8_to_i32_AVX2 and
dot_i8_to_i32_x86.
These do dot[j] = a[i]*b[i] + c[j] where i goes from 0 to 3 and j is the
number of SIMD elements.
a is uint8 and b is int8 and c is int32. In other words it does the dot
product of the 4 bytes in an int for each int32 SIMD element. This is
the same as VNNI.

For dot_i8_to_i32_AVX2 the range of a is limited to [0, 127]. The VNNI
instruction has range [0,255].

The dot_i8_to_i32_x86 which will use VNNI if available otherwise it uses
AVX2. It is only guaranteed to be accurate for a with range [0, 127].

Also adds a
```
 REQUIRES: avx2
```
to test_matmul_vnnni.mojo to fix revert #18110.

LINALG_ORIG_REV_ID: 01516214ea906fb8131c25a99badceb8fa8587b9