[KGEN] Include the headers defining `ssize_t` during header emission (#18170)
When emiting the header file, include the headers the define `ssize_t`.
This has not been an issue up to now because of how we do include
ordering, but we should really not depend on that.

MODULAR_ORIG_COMMIT_REV_ID: 4f9484c5d559a6b0c3622275a2fa07b804552d8d