[Support] Remove footgun in UNWRAP_ERROR_OR_SET usage (#18179)
`UNWRAP_ERROR_OR_SET` consisted of multiple statements, which means you
cannot use it in an if statement without braces. There is a warning
comment on it, but it's not obvious what the problem is when it does
occur. While `UNWRAP_ERROR` also consists of multiple statements, there
is no reason to use it in an if without braces, while for
`UNWRAP_ERROR_OR_SET` there is. Fortunately, since `UNWRAP_ERROR_OR_SET`
is not declaring a new variable the user is expected to use, it is safe
to wrap it in a do-while as is idiomatic, removing the footgun.

MODULAR_ORIG_COMMIT_REV_ID: 903e10f38a60df8856cba7f7087eeeea78d2750d