[All] Generate a modular.cfg file during the build (#18023)
This will allow for modular dev builds to use the same configuration
infrastructure that will be used when releasing to public users and
customers, enabling the removal of a lot of hacks and environment
variables. This PR simply sketches out the basic structure of the
generation, followups will start utilizing this to cleanup the stack.

MODULAR_ORIG_COMMIT_REV_ID: e065bb1e994e240bec9b80e04ba8834748a8eb57