[mojo-tooling] Allow parser listener to specify interest in locations (#18212)
This congregates the logic for determining whether to actually notify
an event on a given location, given in the LSP we only really care about
the current file, and in code completion we only really care about a
specific location. This allows for both removing the need for listeners to
check this, but also lets the parser skip complicated work for an event
on a location if the listener isn't going to care about it.

MODULAR_ORIG_COMMIT_REV_ID: 369441de405db94d64a8710792501b7fc5f111c9