[KGEN] Support different address spaces in KGEN (#18231)
This adds support for specifying address spaces during stack allocation
construction. This would allow us to use things like shared memory (on
GPUs) in Mojo code. Note that the address space enumeration mapping
varies from target to target and the kgen tool is not aware of the
mapping. This mapping will be defined as an enum in the Mojo library.

Note that this does not make changes to the interpreter which still
consideres everything in a global address space (#18230).

Closes #18113
MODULAR_ORIG_COMMIT_REV_ID: 3f97d5ea014f6f6637e3422b2bd2b8002f3b3369