[LLCL] Abstract Resource and ResourceUse for runtime safety checking (#18251)
A Resource represents something than may be referenced, read, or written
to at runtime. A ResourceUse represents one of those active uses. The
Resource tracks its state and active uses and asserts fail if it detects
a use-after-free, data race between concurrent writers and
readers/writers, or a read of uninitialized data. The notion of
'resource section' is also supported.

I'll use this to detect task overhangs and issues with buffers. The
notion of sections is added with an eye to giving runtime protection to
memory-planned models, in which arbitrary sub-sections of the overall
memory pool will come and go during asynchronous execution and it is not
at all trivial to spot if there's a problem just by looking at the MGP
MLIR.

MODULAR_ORIG_COMMIT_REV_ID: 3f8d2f80ccb9cda5dae608da3d922f26ea4b7809