[LLCL][GenericML] Detect 'overhanging' tasks (#18275)
(This is another version of #18106, which I wasn't quite happy with so
didn't submit. But as for #18106, this restores the intent of #18019
which I reverted in #18087 due to false positives. Sorry for all the
churn.)

This adds runtime checking to try to detect when a task launched as part
of a model 'execute' call overhangs the completion of the model. That
is, in a MODULAR_PARANOID build we'll assert fail if, at the point an
await on a model 'execute' call returns, there is some work item
associated with that execute call still in flight.

(The CheckOverhangsPass is the compile-time equivalent to this checking,
and if all is well any model passing that pass will never trigger an
overhang assert.)

- The 'execute' call is represented by an LLCL::Resource
- A referencing LLCL::ResourceUse to that is pushed onto a 'use stack'
before the execute begins.
- Every work item processed by an LLCL::WorkQueue propagates the current
top-of-stack use to any work items it itself schedules.
- Those uses will appear to be alive while the work item is pending, and
while it is executing.
- By default, the use will be automatically ended when the work item
returns all the way back to the LLCL run items loop. However,
WorkQueue::taskIsDone() can be called to indicate the task is done,
which is necessary before emplacing any AsyncValueRef<Chain> which could
directly or indirectly mark the overall 'execute' result chain as ready.
- After the 'execute' await returns, the LLCL::Resource is marked as
'freed', and the system assert fails if any task is still in flight.

I've not added taskIsDone calls throughout the code base, but did wire
them into the Mojo AsyncTaskGroup machinery, which catches the majority
of uses.

MODULAR_ORIG_COMMIT_REV_ID: 033fd4bada31cb772d8916bca5cd097b889980d5