[Stdlib] Implement the M::TensorSpec in Mojo (#18169)
This implements the TensorSpec type that maps to M::TensorSpec in pure
mojo including all the different representations. This is a first step
to providing a Tensor type to Mojo that's compatible with M:Tensor and
would own it's data and be memory managed.

Future PRs would add the docstrings and increase the tests

MODULAR_ORIG_COMMIT_REV_ID: 08851e86114eea8cb418ffa73e0fdb9d247be596