[mojo] Fix mojo-lldb dependency on Windows (#18333)
Fixes a Windows build error introduced by #18319.

On Windows, `LLDB_FOUND` is not true, and so the `mojo-lldb` target
isn't defined, which results in a CMake error when it's added as a
dependency to the `mojo` target. Gate this dependency on `LLDB_FOUND` as
well to resolve the issue.

MODULAR_ORIG_COMMIT_REV_ID: c53b55d79acf38c649823bb7c5bef9d29e788010