[Support] Clean up modular home search paths. (#18365)
This patch formalizes the search paths for the modular home directory. We search the following paths in order:

- $MODULAR_HOME
- $MODULAR_DERIVED_DIR

Then on windows:
- $APPDATA/Local/Modular

Or on unix:
- $HOME/.modular
- /etc/modular

The only interesting bit on Unix is that if $HOME exists, and there's nothing already in /etc/modular, we go back to $HOME/.modular.

MODULAR_ORIG_COMMIT_REV_ID: 324ed20a22886a0937ec91cce4b84634e3b1d664