[MOGG][Kernels] Switch to shape + lambda for kernel api (#18400)
When we take a lambda we should just take a shape directly rather than a
tensor + lambda. It's not legal to load from that tensor so it's weird
to even allow it in user code. Now if the user does not want to take a
lambda they provide an NDBuffer if they do want to take a lambda they
take a shape and a lambda.

Fixes: #18126
Fixes: #16200
LINALG_ORIG_REV_ID: 26c429b4975fd79903b29560c5e717cef9a546aa