[mojo-sdk] Install orcrt as part of the SDK (#18442)
This required supporting a config fallback for the location of the
orc runtime, mojo.orcrt_path, in the same way we do for the compilerrt
location.

This also resolves the orc part of #10097 as part of refactoring to support
config fallback.

MODULAR_ORIG_COMMIT_REV_ID: 9a54f24463e453fc906c306a274b043a42409a6e