[Mojo] Add verifier to `lit.globalvar.decl` (#18498)
This op is similar to `lit.struct.decl` in the sense that it cannot be
nested in other scopes. This patchs endows it with the
`DebugInfo_ScopedLocation` interface and a location verifier. The logic
for both of these is refactored into helpers to reduce code duplication.

MODULAR_ORIG_COMMIT_REV_ID: b17c08d28bde1b932f081a65e26d6f502836d2e8