[Mojo] Integrate Crashpad to mojo driver executable (#18175)
This builds off @yihualou's work in #17384 to add Crashpad to the
Modular mono-repo and get it running for just the Mojo driver for now.
Later, we can add this to more executables as needed, and to the REPL.

It's not clear how to unit test this in its current state. There is no
designed way to force the Mojo driver to crash, and the functionality is
currently limited to the driver. After we support more executables, it
will be more feasible to create a crash-test executable.

Part of #15773.

---------

Co-authored-by: Yihua Lou <1237500+yihualou@users.noreply.github.com>
MODULAR_ORIG_COMMIT_REV_ID: c0fc8d59ece63abc17255976e40062c7e69391b5