[mblack] Bundle mblack as a directory instead of a single executable (#18469)
When mblack is bundled as a single executable, it needs to extract its
contents every single time it's executed, which for some strange reason
takes a few seconds on Mac. A workaround is to bundle mblack as a
directory that doesn't need to be extracted, this results in almost
instantaneous executions of mblack.

I updated the configs for regular modular builds and for the sdk.

MODULAR_ORIG_COMMIT_REV_ID: dd66e04b03b8a0eea816433ecd505e09c312b410