[Telemetry] Initial support for multiple exporters (#18484)
This is a small incremental PR to simply add initial support for
multiple exporters, such that:
- We can export metrics and logs to both a file and an HTTP endpoint.
- We can have separate export destinations for each signal type (for
example: export metrics to file A and logs to file B).

The telemetry config file is likely to change in the future, and so that
part of the code will have to change accordingly.

MODULAR_ORIG_COMMIT_REV_ID: 3914bf7be927ecb960b513e80db07e482b4a52d8