[mojo-sdk] Install jupyter as part of the mojo SDK (#18466)
The jupyter integration is already set up to work outside
of the modular dev environment, so this is simply including
the necessary pieces as part of the cmake install. The kernel
itself will be installed on the user machine as a post install
step.

MODULAR_ORIG_COMMIT_REV_ID: 13c1b6829446230c7fc1d9e1d175f95447e284c7