[vscode-mojo] Strictify the typechecking (#18579)
A couple of times I shot myself on the foot because the typechecker is
not in strict mode, which means that it, for example, doesn't warn on
something being undefined and not being checked for that.

- Add the `strict: true` flag to the typechecker.
- Make all the necessary changes to make the typechecker happy.
- Add two variants of config.get<> to have a sort of intelligent
typechecking on this type.

MODULAR_ORIG_COMMIT_REV_ID: d9d819d7b67fb5536be249a3484cd436f21e5659