[LSP] Index refs to types (#18318)
- This effectively indexes refs to types in decls (argument, variables,
result types, etc.). Fortunately a single event fire location seems to
cover all cases I've tried with. This doesn't include import statements.
- Change SubscriptNode::getLoc() to return the location of the beginning
of the expression. It was returning the location of the opening bracket
[, which seems wrong to me. This might also improve error reporting.
- Teach the LSP to ignore invalid decl locations when creating
onDefinition responses, which I often see from the builtin package. This
allows hovering on types like Bool, showing the documentation, but not
failing when trying to jump to definition, because the decl for this
type has an invalid location.
- Fix the onDefinition result, which was returning the location of the
decl of the symbol, but it was supposed to return the location of the
ref.

MODULAR_ORIG_COMMIT_REV_ID: 9cc70438072b715448445aa1c27ff5bd6388100f