[Cache] Buffer::getFile should not error on an empty file. (#18601)
If a file is empty, Buffer::getFile should not error, it can simply return a BufferRef to an empty string. Because BufferRef is read-only, we don't have to worry about a user writing to it or anything like that.

This also allows us to remove the check in modular-install.cpp because we now handle zero-length files correctly.

MODULAR_ORIG_COMMIT_REV_ID: c019575aae41e2f23b3f9cb29dd621693dfd5529