[Telemetry] Disable OTel internal logs by default (#18531)
OTel internal logs can be enabled with the Modular configuration, with
option `telemetry.internal_log=LOG_LEVEL` otherwise they are off by
default.

The idea comes from Adin's PR:
[Internal link]

In particular, I see that OTel sometimes prints warnings during shutdown
when using the periodic metric exporter (I want to enable the periodic
exporter in a follow-up PR). The issue seems to be that the periodic
exporter worker thread sometimes tries to collect metrics when the
pipeline is shutting down. This warning seems to be unnecessary because
I have never seen this lead to a crash. So it looks like an OTel issue
and in the meantime all we can do is suppress warnings.

Also, OTel prints errors when using the HTTP/OTLP exporter and it can't
connect to the endpoint. It's arguable whether we should suppress this
by default or not, but right now telemetry is not something that users
configure or consume, and if they have outgoing HTTP traffic closed I
don't think they should be seeing this error, or any OTel error for that
matter.

MODULAR_ORIG_COMMIT_REV_ID: 7f3715b66a3b42ba0ac07869454fb930f3b18ffb