[LSP] Don't crash because of CreateClosureOps (#18796)
A CreateClosureOp, which happens with `value_closure` in `decls.mojo`
was making the LSP crash because it was forcefully casting it to a
ASTDeclInterface. Relaxing the cast prevents the crash, but support for
closures needs to be added later anyway.

Closes [Internal link]
MODULAR_ORIG_COMMIT_REV_ID: 3fa2ea595b4e4e379c6d2f0782d28509a7026066