[mojo-lang] Rework magic handling of `def` arguments. (#18789)
Previously we had special handling for `def` arguments that were
complicated to explain and implement.  Now we switch to a simple
model which is that `def` arguments are just "owned" by default,
instead of being "borrowed" by default.  This is easier to
explain and leads to less code in the compiler in aggregate, and
enables a niche use-case mentioned in the changelog.

This fixes #18778 and fixes #13243.

Thank you to Scott for suggesting this!

MODULAR_ORIG_COMMIT_REV_ID: 91871a084f69cc58f3521ca32b50fba54b5a8fcd