[KGEN] Allow one to have global variables without ctor/dtor (#18808)
This allows one to elide the ctor/dtor/priority when declaring global
variables. This is useful for declrations that do not need them (e.g.
they come up when allocating cuda shared memory).

The new syntax is

```
kgen.global @foo : i32 [@foo_ctor, @foo_dtor](2)
```

where the first item in the square bracket is the ctor, the second is
the dtor, and the one in parens is the priority.

The constructors and destructors are optional, and so a global variable
that does not require constructors and destructors can be defined by:

```mlir
kgen.global @my_global : f32
```

MODULAR_ORIG_COMMIT_REV_ID: eb2032c46cb37c99072c752935c534c194334754