[mojo] fix #18599, for loop variable unbound spuriously (#18814)
This fixes issue #18599 where a for loop with an error in the first line
will have spurious undefined variable errors in the body. It fixes the
issue by ensuring that the loop body is always parsed in a context where
the loop variable has a visible ASTDecl.

MODULAR_ORIG_COMMIT_REV_ID: b6e158f694181aeed3767815282904034887026d