[Kernels] VNNI matmul with pre-packed B with any shape (#18849)
With matmul we can pre-pack B (b_packed = True) so that it does not have
to be packed at run-time. This PR implements this for VNNI and emulated
VNNI for any shape of B KxN such as 513x719. Previously the shape had to
have K%4=0 and N%16 = 0.

---------

Co-authored-by: Tracy Sharpe <99041446+raiseirql@users.noreply.github.com>
LINALG_ORIG_REV_ID: 7af29d0e0684a615d653bd639204420ca4300ada