[Support] Implement a custom resource manager for MDialect (#18831)
The resource management needs of the interpreter have some additional
logic that the base resource manager does not implement. For one, this
removes the need for an extra string map to track which resources should
be printed as strings. It isn't particularly important to ensure that
happens but it greatly assists with readability.

This PR is mostly NFC but in the next the ability to deduplicate global
constants will be added, which is particularly important for the string
table.

MODULAR_ORIG_COMMIT_REV_ID: e56db2c80a280147d6d49a08d4b12f5e94381014