[Kernels] remove is_relwithdebinfo checks from tile tuning routines (#18925)
The relwithdebinfo build should be the same as the release build in
terms of optimizations and stack usage, so the code to tune the tile
size should not treat this as a debug build. Remove those checks so that
performance on relwithdebinfo is the same as release.

NN_ORIG_REV_ID: 1030431bb7e831ff168e5c5b97a88a04163e8ed0