[LSP] Add basic crash recovery (#19010)
With this, parsing is executed in a safe environment. Error are reported
as diagnostic notifications attaches to the beginning of the file.
Subsequent keystrokes spawn new parsers, so the experience doesn't get
too bad.

I want to implement a proper popup notification for this besides the
diagnostic, and I plan to use the `data` field of `Diagnostic` from the
protocol to send this additional piece of information, which could be
captured by the middleware, but this field hasn't been implement in the
mlir structs yet. I'll make the change in LLVM and after that I'll
implement the notification.
The other option is to implement a new notification kind, but
vscodelc.LanguageClient doesn't support that, so we might need to switch
to other language client, which might be an overkill.

MODULAR_ORIG_COMMIT_REV_ID: a43724209ce07762adf50d59499bd8fdf5fbf47c