[Stdlib] Move files around in anticipation of the Math package (#19019)
This just moves files into a different folder to anticipate creating the
math package. There are a few things that are done though:

- this is just a reorg, it does not change the fact that the math folder
is still shipped as source
- the folder name is now lower case. This is to allow us to do `import
math` instead of `import Math`. The former is what folks in Python are
used to. The capitalization was an artifact of writing mojo code in
January
- there are a few issues with imports that @River707 will resolve and
based on this package the math folder

LINALG_ORIG_REV_ID: c12b0d6298d6b57a413b4d34a2929eed39cdc276