[Support][modular-tool] Fix llvm::formatv usage. (#19100)
It seems like llvm::formatv changed in the last integrate, and now
passing the result into something that takes a Twine results in the use
of a temporary that is removed - this is undefined behavior. This patch
explicitly constructs an std::string so that we avoid that behavior.

MODULAR_ORIG_COMMIT_REV_ID: 37c4b32f065e27b59279a6fbe10bd56f91385f8c