[Stdlib] Correctly propagate the address space for CUDA allocations (#18717)
This is ugly since it requires a copy/paste of code. But, at the
moment we have to copy/paste the code. To be able to delete the
code we need a combination of things, but the main one is that
we need to have keyword parameters so that we can write
Pointer[type, address_space=0] that would allow us to unify the stack
allocations that take an address space (defined here) with the ones that
do not take an address space (defined in Memory.mojo).

Closes #18714
MODULAR_ORIG_COMMIT_REV_ID: e5ff6b2394e40bc540c2429fb4962e29ab32112a