[Support] Remove padding from URL-safe Base64 encodings. (#19141)
This patch removes the padding from the URL-safe Base64 encodings. According to the standard, it's optional, and it requires escaping in URLs, so we just remove it on encode and add it back in on decode.

This also means we have to regenerate the modular tool tests because the filesystem backend looks for url-safe encoded filenames, now without padding at the end.

MODULAR_ORIG_COMMIT_REV_ID: fa83259e7218bbba37a4c650657c827d6c203d2d