[Mojo] Reject import with the wrong case even on case-insensitive filesystems (#19159)
Some file systems (like Windows and MacOS) are case insensitive. The
import mechanism we had would follow the same sensitivity rules as the
file system -- resulting in behavior that's platform specific. This
changes that so that module imports honor the case sensitivity.

Currently, this is a hard error, but we can relax that and give better
diagnostics to the user and/or follow what clang does and give a
warning.

Closes #19081
MODULAR_ORIG_COMMIT_REV_ID: 679ca802798d9520da5b363272aa2e4e1135a96a