[mojo-stdlib] Turn the math folder into a source package (#19062)
The contents of the math module are re-exported into the package.
A followup will handle turning this into a binary package, but
this PR handles updating all of the references to the old modules to
import via the math package.

MODULAR_ORIG_COMMIT_REV_ID: 6faa34fa08363cacd73ff6d3bf1faeab734d1827