[Support][modular-tool] Refactor file finding. (#19185)
We may have files in a few places on the filesystem (especially coming from the native package for the installer), and we want that set to be a superset of the list of paths that we could use as an install home.

This patch sets up search paths and utilities for finding files using a standard set of search paths. Specifically, we can find files in /etc/modular on unix platforms, but we don't want to use /etc/modular as an install base.

MODULAR_ORIG_COMMIT_REV_ID: 373b7683ac7e5a6cbd4b33d1a6b529bbe62c10d1