[KGEN] Add interpreter support for `pop.compiler.global_*` ops (#19189)
Shoutout to Lukas for complaining about this.

This PR adds support for interpreting functions with parameter closures
by teaching the interpreter how to handle the `pop.compiler.global_load`
and `pop.compiler.global_store` ops. Note that KGEN's handling of
parameter
closures is incorrect due to #19175. The correct way to handle this in
the interpreter is to store the "named globals" map on the stack frame,
thereby giving scoped identities to the SSA values. However, this PR
does not do that in order to keep the buggy behaviour consistent.

MODULAR_ORIG_COMMIT_REV_ID: c13eae125d36c490a93e93b2ad58caa840551702