[Support][DebugInfo] Correct location when lowering `debuginfo.value` for arguments (#19178)
The location of the `llvm.store` op for the debuginfo value cannot
directly use the location of block arguments, since these are not
required to carry debuginfo scopes. Instead, we can simply use the
location of the parent op. In the future, we might improve this if/when
we have verifiers for block argument location info, but for now this
unblocks a kernel test (enabled with debuginfo in this patch).

MODULAR_ORIG_COMMIT_REV_ID: fc9ad36ce0f57536bc246f83e0a72b62d35b5dbc