[triton] Add thread names to enable debugging (#18147)
This PR adds a `poolName` parameter to the thread pool queues which allow the creator of the runtime pool to specify a pool name which will be used to name the threads created in the pool.

The runtime config initialization checks the modular.cfg file for a thread identifier to use, if specified, under the section 'triton_modular_backend' and key 'context_name'. If found, use it as the context name, otherwise use 'LLCL Thread ' as the prefix for threads in the thread pool.

MODULAR_ORIG_COMMIT_REV_ID: b6abbb84e8b5d2948273a02e40aea53ceaeb04f1