[Support] [Crashpad] Mark tests as unsupported on Windows where applicable (#19241)
One test (`default-find-handler.mlir`) can run on Windows as-is. All
other tests here rely on POSIX shell functionality and cannot run on
Windows as-is. The functionality in the Crashpad is present in Windows
and is theoretically testable, but I don't know how to write it in a way
that runs on Windows. I filed #19240 to create Windows equivalents of
these tests.

MODULAR_ORIG_COMMIT_REV_ID: 5e02ff23218290f065a8bfe161b1945f0e92cecc