[Support] [Crashpad] Detect main executable when searching for handler (#19229)
Previously we relied on:

1. Configuration
2. argv[0]
3. PATH.

But, under mojo-lldb in CI (and potentially other scenarios), none of
the Modular tools are on the PATH (so 3 fails), we don't have argv[0]
since we're only an LLDB plugin (so 2 fails), and nothing was specially
configured to override (so 1 fails), so we can't find the handler.

We insert an additional test, attempting to use
llvm::sys::fs::getMainExecutable between 2 and 3. This should handle the
case that mojo-lldb finds itself in today.

MODULAR_ORIG_COMMIT_REV_ID: ee6aedb861ce0589bcb109653f39df20d510bf29