[Stdlib] Move DType to be a builtin type (#19247)
This moves DType to be a builtin type instead of being a DType module.
The DType underlies a ton of our functionality and so it deserves to be
a builtin type. Futheremore, we are moving SIMD to be a builtin type,
and so it makes sense for DType to follow suite since SIMD uses DType

thanks to @River707 for fixing my typos

MATH_ORIG_REV_ID: 0582342b8a8dba476b33372c6049626c1a5c94ac