[mojo-lang] Refactor TokenError emission to return InFlightDiag. (#19270)
This changes the Mojo Lexer's emitTokenError method to return an
InflightDiag, allowing us to attach notes to it.  This is mostly
a mechanical patch, with an improvement that helps compiler debugging
of a patch that is coming up next.

MODULAR_ORIG_COMMIT_REV_ID: 0a5e728b6405aae99add3918494c87598dcd7332