[mojo-driver] Print GH issues URL for LLVM crashes (#19292)
When LLVM crashes, instead of printing the default error message (which
asks users to report to https://github.com/llvm/llvm-project/issues),
print one that points users to the Mojo GitHub page.

This was an oversight when bringing up the `mojo` driver -- for tools
such as `kgen`, we print this same message but pointing at our internal
employee URL [Internal link] I forgot to
specify a custom LLVM crash message for the `mojo` driver.

As for which URL to use: as of today we have 4-5 people dogfooding our
private beta, and they have been told to send bug reports to
mojo-beta-feedback@modular.com. However, once the SDK becomes available
as an external closed beta on August 21 2023, we'll no longer have to be
so secretive, and users will be able to report to
[Internal link] So, instead of changing this
message to use the email address today, and then making another change
to have it use the issues link in a week, just change it to the link
now. We will tell our external and employee dogfooders that, no matter
what this message says, just report via email or the internal employee
URL.

Closes #19288.
MODULAR_ORIG_COMMIT_REV_ID: 0106d7663a26a9d8d3d374a9230a87750e01e18e