[mojo] Disable "validate all docs" test (#19340)
As originally written, this test was meant to validate docs only present
in the source file itself. A comment in the test says as much:

```
 """This definition of `object` is required because the compiler performs a
 lookup of an "object" type for `def` functions. We cannot
 `from Object import object` because doing so would `-doc-validate` the
 `Object.mojo` file and all its dependencies."""
```

However, over time the Mojo compiler changed to import more and more
code. For example, AFAICT currently every Mojo program imports DType,
which in turn imports modules like LLCL.

Because `mojo doc -validate` validates all doc strings, including ones
that appear in transitively imported code, this results in this one test
validating pretty much every doc string in the entire stdlib. However,
not all modules should be validated. LLCL, for example, is not meant to
ever be exposed to users, and so does not require extensive public
documentation.

Because of the transitive imports, this test should have failed 100% of
the time. And it would have, were it not for another quirk: doc strings
in cached modules are not validated. This allowed for a workaround to be
added to the test in #18905 that would allow the test to sometimes pass.

This commit removes the cache-based workaround and disables the test, at
least until the Mojo parser is modified to not validate doc strings in
imported files.

Addresses a portion of #18903.

MODULAR_ORIG_COMMIT_REV_ID: e0a7c556a1d607aa7d5e3cf80c3d2ba238132d64