[MOGG] Add support to treat range as an elementwise op (#19239)
This allows for range to be supported as an elementwise op. This allows
it to become a trivial fused kernel with slightly different semantics to
the others. As it takes in tensors instead of SIMD arguments it will not
fuse into those tensors but will still allow itself to be fused through
the SIMD return value (output fusion).

MOGG_ORIG_REV_ID: c2a8bd3a58a8dd8042ba3b75f93ca48c649b26ba