[mojo] Don't validate docs in imports (#19437)
Closes #18903.

Instead of validating doc strings in all parsed Mojo code, only validate
doc strings in the "input," meaning either a Mojo source file or a Mojo
package -- but NOT the files or packages that that input imports.

This allows for the disabled `parser-doc-errors.mojo` test to be
re-enabled, and for the simplification of the `modular-doc-validate`
build target (which now no longer needs to use a regular expression to
filter warnings to only ones emitted from locations in the input file).

The new timing of the doc validation, in
`DeclResolver::resolveAllReferencedFrom`, means that decls that have
been synthesized (such as initializers synthesized for structs decorated
with `@value`) are now potentially validated. To avoid unwanted
diagnostics for functions that don't actually appear in source code, use
the assumption that synthesized methods will always have the same
location as their parent struct.

MODULAR_ORIG_COMMIT_REV_ID: 7f0a3a2fc5dc51c5460221276aed796133adae61