[Support] [Crashpad] Add default Crashpad upload URL (#19369)
This sets the default Crashpad URL to the API Gateway that Kevin set up
in front of the Lambda.

The Lambda currently requires an `auth` parameter to be provided, but as
far as I could tell, Crashpad did not have a way to provide this
parameter, so the upload will fail until either we find a way to provide
it or the Lambda is changed to no longer require this parameter.

MODULAR_ORIG_COMMIT_REV_ID: d60ffe51c4677ff905d6525c956fed9e809a1107