[mojo-jupyter] Allow enable_matplotlib to fail gracefully (#19375)
At the start of the jupyter kernel we run a python expression to
initialize matplotlib and define a backend that displays images
inline. This PR allows for that initialization to fail and be ignored,
as we shouldn't blow up the jupyter kernel if, for example, the
user doesn't have matplotlib installed.

MODULAR_ORIG_COMMIT_REV_ID: ab8d038e52e505996a232a9571b98af3c9ee5030