[KGEN] `get_all_impls` fails if there are no valid candidates (#19515)
This parameter operator was allowed to return zero candidates, but in
current uses cases that is not particularly useful and it results in an
opaque out-of-bounds variadic get error. It's hard to pin down the
source because this effectively swallows the original elaborator errors.

Instead, just fail and emit the usual elaborator errors if there are no
successful instantiations.

MODULAR_ORIG_COMMIT_REV_ID: 40940dca61ef304be023b440cc027bac38c0bf70