[mojo] Remove `mojo -L` compilation option (#19619)
Remove the `-L` option from the `mojo` driver and its subcommands. This
option was meant to add paths to the ones `mojo` searched for its
runtime libraries, such as KGEN Compiler RT. But we now find those
via the `MODULAR_HOME` environment variable and the `modular.cfg`
configuration file. Since we don't actually use it ourselves in the
`mojo` driver, and since users especially have no use for it (they have
no means of integrating with the Mojo runtime), remove the option
altogether.

Note that this commit doesn't remove the `kgen -L` option, which is used
in a small number of tests. We may wish to remove that also, but such a
change would not be user-facing.

Addresses a portion of #19459.

MODULAR_ORIG_COMMIT_REV_ID: 50b8049b57766a96a50a388ac9050fd491ac0a67