[driver-tblgen] Option help text ends in `.`, `"` (#19623)
Enforce that command line options' help text must end in either a period
`.` or quotation mark `"`, and fix the handful of occurrences in the
`mojo` driver where this was not the case.

Addresses a portion of #19459.

MODULAR_ORIG_COMMIT_REV_ID: 31dd36ae5f67acc4461ce6ef57b3a60384b08093