[Cache] Switch size/offset in to optionals. (#19642)
Buffer::getFile has size and offset arguments that are optional arguments. The implicit assumption originally was that the file being read would not be zero-sized. This has proven to be incorrect, especially when sub-setting a file with a computed size. This patch changes the size_t parameters to be std::optional<size_t> so that we can more easily tell if it's "unspecified" or "zero".

MODULAR_ORIG_COMMIT_REV_ID: dfa692b883e67a560adbe4f2d4d54927d18d7815