[driver-tblgen] Remove automatic `man` invocation (#19850)
Replace the current `--help` option, which implicitly invokes `man` on
host machines that have it in the `PATH`, with the behavior from
`--help-text`, which just prints help text to stdout.

Fixes #18962.

This doesn't change any of the assertions made in tests that verify the
output of `--help`. Many of these tests are brittle and their
assertions should be reduced significantly (there's no need to test the
exact text that is printed in help text, and doing so makes it difficult
for that text to be copy-edited), but this will be done in a follow-up
commit.

MODULAR_ORIG_COMMIT_REV_ID: f8d36e250f2dbf4666f296241896d5d13b8f9005