[mojo-sdk][modular-tool] Don't test help text (#19855)
There's no need to test the exact text that is printed in help text, and
doing so makes it difficult for that text to be copy-edited. Remove
assertions about the specific text printed in favor of tests of the
behavior of the `--help` option (whether it supercedes invalid
arguments, etc.).

In the process of modifying these tests, I discovered a bug, which this
patch also fixes. Previously, invoking `modular --invalid-argument`
would do nothing and return a successful exit code:

```
$ modular --invalid-argument
```

This patch turns this into an error:

```
$ modular --invalid-argument
modular: error: unrecognized argument '--invalid-argument'
```

If the user specifies `--help` on the command line, however, help text
is printed and a successful exit code is returned:

```
$ modular --invalid --help
```

MODULAR_ORIG_COMMIT_REV_ID: af37983524d34dd33314f9048a948d66b2553a73