[mojo-sdk] Remove `name` argument from mojo package (#19864)
This already needs to be the same as the output, so just infer it
from the output file name, or the input package when outputting
to stdout. This also adds a proper check that the output file ends
with the expected file extensions.

MODULAR_ORIG_COMMIT_REV_ID: 37deb4715b20921b353b225a3464a15d24aa1f5d