[mojo-lang] Add PValue to MValue conversions (#19797)
And fix a couple of other issues with memory-primary values. This
completes the piping in of memory-primary parameter values. There is a
minor hack in that default values have to match the type of the
arguments, requiring an extra wrapper attribute, but that should
disappear with time.

Fixes [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: c6d06514be3d86446488b930e877c43258e30df2