[KGEN] Add a `kgen.param.materialize` operation (#19990)
This operation plays the same role as `kgen.param.constant`, except that
it is side-effecting (MemAlloc on the result). This operation should be
used when the parameter expression may have memory allocations. It
decays to `kgen.param.constant` when the RHS expression is a simple
constant and does not contain memory references.

MODULAR_ORIG_COMMIT_REV_ID: 0c5eae030ec0632852752bff578b63592e455bdb