[KGEN] Exclude Escaping Closures From Explicit Copy (#19372)
In generating the __call__ method I removed the copying of the captures
since the copy will occur before the initialization of the closure. That
resulted in an order change in the fields since the copies were
resulting in reverse order occurrences of the captures. This diff fixes
the test after the exclusion.

MODULAR_ORIG_COMMIT_REV_ID: 3d6a166ece066acd97cfffb801a423c5b6ed7f02