[mojo-sdk] Don't load `.lldbinit` in REPL tests (#20066)
The immediate motivation for this change is to fix an issue when running
post-`modular install` self-tests from a directory that contains an
`.lldbinit` file. The REPL self-test checks that stderr is empty, but
lldb's default behavior is to print the following warning to stderr when
an `.lldbinit` file exists in the current working directory:

```
$ mojo
error: There is a .lldbinit file in the current directory which is not being read.
To silence this warning without sourcing in the local .lldbinit,
add the following to the lldbinit file in your home directory:
    settings set target.load-cwd-lldbinit false
To allow lldb to source .lldbinit files in the current working directory,
set the value of this variable to true.  Only do so if you understand and
accept the security risk.
Welcome to Mojo! 🔥
Expressions are delimited by a blank line.
Type `:mojo help` for further assistance.
  1>
```

This unexpected stderr output causes the self-tests to fail when run in
a directory that contains an `.lldbinit` file.

More broadly, however, we may wish for the Mojo REPL to no longer expose
its integration with lldb in this way: the fact that the REPL is built
on top of lldb is a technical detail, not necessarily a feature. In that
respect, loading `.lldbinit` files is a feature of lldb, not of the Mojo
REPL (if anything, our users would expect to load files named
`.mojorepl` or something). -- and so this feature maybe ought to be
disabled in `mojo repl` itself using the `lldb --no-lldbinit` option.
But if so, that'll happen in a future commit.

MODULAR_ORIG_COMMIT_REV_ID: 4d5643f650ddd49d740efb0fb2928723c2f06c77